fix: internal refactoring to reduce file size when bundled
