+++ lib/schemas/validate.js
@@ -1,7 +1,7 @@
 import { readFileSync } from "node:fs";
 import { join, dirname } from "node:path";
 import formats from "ajv-formats";
-import semver from "semver";
+import validSemver from "semver/functions/valid.js";
 import npmPkg from "validate-npm-package-name";
 import Ajv from "ajv";
 import { fileURLToPath } from "node:url";
@@ -88,7 +88,7 @@ const createNameValidator = (jsonSchemaValidator) => (value) => {
 const createVersionValidator = (jsonSchemaValidator) => (value) => {
 	const result = jsonSchemaValidator(value);
 	if (!result.error) {
-		const version = semver.valid(value);
+		const version = validSemver(value);
 		/** @type {ErrorObject[]} */
 		const errors = [];
 		if (!version) {
