+++ lib/validators/extra.js
@@ -1,0 +1,8 @@
+/**
+ * No-op, returning its value.
+ * @param {string} value
+ * @returns {string} the value
+ * @todo https://github.com/asset-pipe/core/issues/12
+ * @todo Can we remove this?
+ */
+export const extra = (value) => value;
